Complete the iPad app: editor coordinator, library, review, search, export and settings
Editor: a NotebookEditorViewController that owns one undo manager shared
with every PencilKit canvas, virtualizes pages to three live canvases,
commits debounced ink as one undoable operation with its object edits,
drives save status and page navigation, and a SwiftUI entry point with
reading mode and in-notebook search. Shell: the app entry point and
environment, a procedural design system, the library with folders,
courses, trash and import, the course review queue, the Problem Page
inspector, search with indexing state, export and backup, settings and
onboarding. Adds simulator tests for the ink engine, page layout, tool
state, notebook search, app shell, PDF export alignment, tape policy and
text searchability.

219 Linux tests pass; 74 simulator test functions await the CI run.

+++ App/Courseleaf/Editor/PageRenderer.swift
@@ -102,7 +102,7 @@ enum PageRenderer {
             let a = point(shape.start), b = point(shape.end)
             path.move(to: a); path.addLine(to: b)
             let length = hypot(b.x - a.x, b.y - a.y)
-            let head = min(max(shape.strokeWidth * 5, 10), max(length / 2, 1))
+            let head = min(max(CGFloat(shape.strokeWidth) * 5, 10), max(length / 2, 1))
             if length > 0.5 {
                 let angle = atan2(b.y - a.y, b.x - a.x)
                 let left = CGPoint(x: b.x - head * cos(angle - .pi / 6), y: b.y - head * sin(angle - .pi / 6))
